Updated to API 34 (#74)
The Android app needed to be updated to use API 34 so that it meets the App Store requirements. In the process, this also updates the gradle version, the Robolectric version, the ruby gem versions, and the npm package versions. It also changes the lint workflow to use java 17 and fixes a bunch of build warnings and errors caused by deprecated methods, unused variables, and outdated libraries.