Fix release build (#76)
This changes the build to use Java 17 and fixes a release build simulation mode crash.
In release builds, simulation mode would crash due to the MQTT manager failing to find the logging class. The solution appears to be adding the appropriate classes to proguard so that they don't get obfuscated or renamed.